[♻️refactor]: api 호출 코드 분리 리팩토링
